web/requirements.txt
flask==2.0.3
werkzeug==2.0.3
kafka-python==2.0.2
